gtv holds the WIP of what gtktv was supposed to be - a GtkTV widget. tgtv
gtv holds the WIP of what gtktv was supposed to be - a GtkTV widget. tgtv
demos it.

If anyone wants to help out, a very good place to start would be handling
the GDK_VISIBILITY_PARTIAL case by generating the clipping rectangles from
the results of an XQueryTree() (there is no better way).