Massive changes to TCD's cddb code. The actual socket code has been moved from the application to a slave program called cddbslave. This allows the client (tcd) to call the slave with the request, and then immediatly return control back to the user. The slave sends a SIGUSR1 to the client when the request has been finished, signalling the client to reload the info and redisplay.
TODO:

The slave needs to cache requests so that a batch of discs can be identified
at a time, for those lowly dialup users. (myself included)

HTTP/Proxy support.

We need a global configuration program for the slave, to set servers and so
on. Right now it's hardcoded to cddb.cddb.com:888.

Some way of communicating status back to the client would be nice...shm?