Add support for new GStreamer mixer flags
Enhancement request #570832 added new flags to gst-plugins-base 0.10.23.  This
commit adds support for these new flags to gst-mixer.  These new flags allow
options and switches to be displayed on the Record and Playback tabs, allow
tracks to specify that they do not support mute, and to control whitelist
behavior.  This change is backwards compatible, and plugins which do not
support these flags will work as always.  Also configure.ac is updated to
specify that GStreamer 0.0.10.23 is now needed when building gst-mixer.
(Bug #571106)