Fix stream volume vs. muted status problems
We need to behave differently when we have and don't have a muted
button in the channel bar.

Only mute and unmute the channel bar on 0-level when we don't
show a mute button (status icon)

Tell front-ends of the value change when we switch between
the zero adjustment, and the normal one.

Don't unmute muted streams when raising the volume (dialogue)

Set the stream mute status before changing it in PA.

All this fixes the status icon's slider not unmuting streams, and
staying stuck to zero.

https://bugzilla.gnome.org/show_bug.cgi?id=595857