Force updating the custom theme dir on change
When changing only the alert sound, changing the links inside
the __custom theme will not update the directory's mtime, thus
not trigger gnome-settings-daemon's sound module to flush
the sound cache, and the old sound would then be used.

Poke at the __custom theme dir when changing the alert sound
fixes this.

https://bugzilla.gnome.org/show_bug.cgi?id=620360