Port to GTK+ 3.0
This includes a rename of the libgnome-media-profiles.so library (renamed to
libgnome-media-profiles-3.0.so); but the issue of a shared gconf schema
is not solved, and the fix will probably be in a GSettings port later.

Also it depends on an unrealeased gladeui-3.0 library to register the profile
widget in the glade3 catalog.