grecord: send eos before we stop record.
Current grecord produces files without EOS. This can be tested
with tool "ogginfo".

accodrding to this documentation:
http://gstreamer.freedesktop.org/data/doc/gstreamer/head/gstreamer-libs/html/GstBaseSrc.html
"Controlled shutdown of live sources in applications"
"Since GStreamer 0.10.16 an application may send an EOS event to a source element to make it perform the EOS logic".

So this patch make sure we send EOS before we stop recording.

Signed-off-by: Oleksij Rempel <bug-track@fisher-privat.net>

https://bugzilla.gnome.org/show_bug.cgi?id=677644